+++ TabletopTweaks-Reworks/ModLogic/ModContextTTTReworks.cs
@@ -20,8 +20,8 @@ namespace TabletopTweaks.Reworks.ModLogic {
         public override void AfterBlueprintCachePatches() {
             base.AfterBlueprintCachePatches();
             if (Debug) {
-                //Blueprints.RemoveUnused();
-                //SaveSettings(BlueprintsFile, Blueprints);
+                Blueprints.RemoveUnused();
+                SaveSettings(BlueprintsFile, Blueprints);
                 ModLocalizationPack.RemoveUnused();
                 SaveLocalization(ModLocalizationPack);
             }