+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Version 1.1.1
 * Bugfixes
 	* Fixed issue with some trickster domain powers granting bonuses of the wrong type.
+* Lich
+	* Fixed Fear Control to use new owlcat DC logic.
 
 ## Version 1.1.0
 * Mythic Abilities